fix(templating): fix text position in area helper when collapse=false
Fixes positioning of the text content in area helper with collapse=false option. Before, the emitted block was

```
#- {area="test"}
content
#- {area_end="test"}
```

which means that the text would be part of the area start block. In this commit, the output was changed to

```
#- {area="test"}
#-
content
#- {area_end="test"}
```

so that the content will always start in its own block